moved commons under table folder
