+++ azure-cosmosdb-table/azure/cosmosdb/table/common/_auth.py
