+++ azure-cosmosdb-table/azure/cosmosdb/table/common/_connection.py
