+++ azure-cosmosdb-table/azure/cosmosdb/table/common/_constants.py
