+++ azure-cosmosdb-table/azure/cosmosdb/table/common/_deserialization.py
