+++ azure-cosmosdb-table/azure/cosmosdb/table/common/_encryption.py
