+++ azure-cosmosdb-table/azure/cosmosdb/table/common/_error.py
