+++ azure-cosmosdb-table/azure/cosmosdb/table/common/_http/__init__.py
