+++ azure-cosmosdb-table/azure/cosmosdb/table/common/_http/httpclient.py
