+++ azure-cosmosdb-table/azure/cosmosdb/table/common/_serialization.py
