+++ azure-cosmosdb-table/azure/cosmosdb/table/common/cloudstorageaccount.py
