+++ azure-cosmosdb-table/azure/cosmosdb/table/common/models.py
