+++ azure-cosmosdb-table/azure/cosmosdb/table/common/retry.py
