+++ azure-cosmosdb-table/azure/cosmosdb/table/common/sharedaccesssignature.py
