+++ azure-cosmosdb-table/azure/cosmosdb/table/common/storageclient.py
