+++ azure-cosmosdb-table/azure/cosmosdb/table/common/tokencredential.py
