+++ azure-cosmosdb-table/samples/test_sample.py
@@ -14,7 +14,7 @@
 # --------------------------------------------------------------------------
 import unittest
 
-from azure.cosmosdb.common import CloudStorageAccount
+from azure.cosmosdb.table.common import CloudStorageAccount
 from samples.advanced import (
     AuthenticationSamples,
     ClientSamples,
