+++ azure-cosmosdb-table/tests/table/test_table_entity.py
@@ -26,10 +26,10 @@ from azure.common import (
 )
 from dateutil.tz import tzutc, tzoffset
 
-from azure.cosmosdb.common import (
+from azure.cosmosdb.table.common import (
     AccessPolicy,
 )
-from azure.cosmosdb.common._common_conversion import (
+from azure.cosmosdb.table.common._common_conversion import (
     _encode_base64,
 )
 from azure.cosmosdb.table import (
